Change structure, separate by file types
