+++ .config/telex2reddit.logging.json
@@ -1,0 +1,151 @@
+{
+	"version": 1,
+	"filters": {
+		"notset_level_filter": {
+			"()": "telex2reddit.levelfilter.NotsetLevelFilter"
+		},
+		"debug_level_filter": {
+			"()": "telex2reddit.levelfilter.DebugLevelFilter"
+		},
+		"info_level_filter": {
+			"()": "telex2reddit.levelfilter.InfoLevelFilter"
+		},
+		"warning_level_filter": {
+			"()": "telex2reddit.levelfilter.WarningLevelFilter"
+		},
+		"error_level_filter": {
+			"()": "telex2reddit.levelfilter.ErrorLevelFilter"
+		},
+		"critical_level_filter": {
+			"()": "telex2reddit.levelfilter.CriticalLevelFilter"
+		}
+	},
+	"formatters": {
+		"colorless_formatter": {
+			"format": "[{asctime}] {message}",
+			"datefmt": "%H:%M:%S",
+			"style": "{"
+		},
+		"grey_formatter": {
+			"format": "\u001b[30m[{asctime}] {message}\u001b[0m",
+			"datefmt": "%H:%M:%S",
+			"style": "{"
+		},
+		"red_formatter": {
+			"format": "\u001b[31m[{asctime}] {message}\u001b[0m",
+			"datefmt": "%H:%M:%S",
+			"style": "{"
+		},
+		"white_formatter": {
+			"format": "\u001b[37m[{asctime}] {message}\u001b[0m",
+			"datefmt": "%H:%M:%S",
+			"style": "{"
+		},
+		"reverse_red_formatter": {
+			"format": "\u001b[7m\u001b[31m[{asctime}] {message}\u001b[0m",
+			"datefmt": "%H:%M:%S",
+			"style": "{"
+		},
+		"reverse_magenta_formatter": {
+			"format": "\u001b[7m\u001b[35m[{asctime}] {message}\u001b[0m",
+			"datefmt": "%H:%M:%S",
+			"style": "{"
+		},
+		"complex": {
+			"format": "[{asctime}]{levelname}: {message}",
+			"datefmt": "%H:%M:%S",
+			"style": "{"
+		}
+	},
+	"handlers": {
+		"colored_notset_handler": {
+			"class": "logging.StreamHandler",
+			"level": "NOTSET",
+			"filters": [
+				"notset_level_filter"
+			],
+			"formatter": "colorless_formatter"
+		},
+		"colored_debug_handler": {
+			"class": "logging.StreamHandler",
+			"level": "DEBUG",
+			"filters": [
+				"debug_level_filter"
+			],
+			"formatter": "white_formatter"
+		},
+		"colored_info_handler": {
+			"class": "logging.StreamHandler",
+			"level": "INFO",
+			"filters": [
+				"info_level_filter"
+			],
+			"formatter": "grey_formatter"
+		},
+		"colored_warning_handler": {
+			"class": "logging.StreamHandler",
+			"level": "WARNING",
+			"filters": [
+				"warning_level_filter"
+			],
+			"formatter": "red_formatter"
+		},
+		"colored_error_handler": {
+			"class": "logging.StreamHandler",
+			"level": "ERROR",
+			"filters": [
+				"error_level_filter"
+			],
+			"formatter": "reverse_red_formatter"
+		},
+		"colored_critical_handler": {
+			"class": "logging.StreamHandler",
+			"level": "CRITICAL",
+			"filters": [
+				"critical_level_filter"
+			],
+			"formatter": "reverse_magenta_formatter"
+		},
+		"file_info": {
+			"backupCount": 30,
+			"class": "logging.handlers.TimedRotatingFileHandler",
+			"encoding": "utf-8",
+			"filename": "../log/info.txt",
+			"formatter": "complex",
+			"level": "INFO",
+			"when": "midnight"
+		},
+		"file_error": {
+			"backupCount": 52,
+			"class": "logging.handlers.TimedRotatingFileHandler",
+			"encoding": "utf-8",
+			"filename": "../log/error.txt",
+			"formatter": "complex",
+			"level": "ERROR",
+			"when": "W0"
+		}
+	},
+	"loggers": {
+		"root": {
+			"handlers": [
+				"colored_notset_handler",
+				"colored_debug_handler",
+				"colored_info_handler",
+				"colored_warning_handler",
+				"colored_error_handler",
+				"colored_critical_handler",
+				"file_info",
+				"file_error"
+			],
+			"level": 0
+		}
+	},
+	"rollbar": {
+		"enabled": true,
+		"handler": {
+			"access_token": "",
+			"environment": "production",
+			"level": "WARNING"
+		}
+	}
+}
