+++ Makefile
@@ -1,5 +1,8 @@
+init:
+	pip install -r requirements.txt
+
 docker-start:
-	docker run --name telex2reddit --restart unless-stopped -d -v "`pwd`:/srv" rutkai/telex2reddit
+	docker run --name telex2reddit --restart unless-stopped -d -v "`pwd`:/srv" -e HOME=/srv rutkai/telex2reddit
 
 docker-build:
 	docker build . -t telex2reddit
