--- telex2reddit.ini
@@ -1,44 +1,0 @@
-[reddit]
-english_collection_id=3c4f22b5-4354-4af7-a3c8-d0f66ad9414e
-script_author=ForroKulcs
-subreddit=telex
-subreddit_english=hungariannews
-username=telex_bot
-
-[collect_links]
-english=https://telex.hu/english
-frontpage=https://telex.hu
-kedvenceink=https://telex.hu/kedvenceink
-kozlemenyek=https://telex.hu/kozlemenyek
-legfrissebb=https://telex.hu/legfrissebb
-legfontosabb=https://telex.hu/legfontosabb
-
-[telex]
-api_url=https://telex.hu/api/articles
-articles_per_page=25
-article_cache_valid_time=86400
-check_interval=300
-use_article_cache=1
-useragent=Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:81.0) Gecko/20100101 Firefox/81.0
-
-[categories]
-belfold=Belföld
-chart=Chart
-defacto=Defacto
-english=English
-eszkombajn=Észkombájn
-foto=Fotó
-gazdasag=Gazdaság
-interju=Interjú
-koronavirus=Koronavírus
-kulfold=Külföld
-kult=Kult
-napirajz=Napirajz
-nevertek=Névérték
-sport=Sport
-szorakozas=Szórakozás
-tamogatott-tartalom=Támogatott Tartalom
-tech=Tech
-tudomany=Tudomány
-velemeny=Vélemény
-video=Videó
