+++ telex2reddit/__init__.py
@@ -1,0 +1,1 @@
+from .telex2reddit import init, main
